Remove in-body inline JS where possible for our reCAPTCHA.
Both recaptcha types have wp_add_inline_script options now and we need to enqueue both.
So, instead of doing a big if/else statement, we will grab the current saved version, and new up a class based on that.
Then we call the enqueue scripts method we know will be available.