Due to the fact that I can't find anything in the current codebase that refers to a "connection-settings-redirect" html class, I think this small part of the javascript is orphaned.
On top of that, I checked out the 2.0.0 branch when it was first introduced, and still couldn't find any references to the class. Given the scope and size of the 2.0.0 release, I
must conclude that the javascript spot was originally implemented at one time during the 2.0.0 release work, but a pivot was made and the js was missed for removal afterwards.

I have tested the connection process without this snippet in the build files, and our workflow remained unaffected. If by chance we need to re-add in, after a release and support
comes in, we most definitely can.